+++ packages/bluesky/lib/src/services/entities/actor_viewer.freezed.dart
@@ -32,9 +32,14 @@ mixin _$ActorViewer {
   AtUri? get following => throw _privateConstructorUsedError;
   @AtUriConverter()
   AtUri? get followedBy => throw _privateConstructorUsedError;
+  KnownFollowers? get knownFollowers => throw _privateConstructorUsedError;
 
+  /// Serializes this ActorViewer to a JSON map.
   Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
-  @JsonKey(ignore: true)
+
+  /// Create a copy of ActorViewer
+  /// with the given fields replaced by the non-null parameter values.
+  @JsonKey(includeFromJson: false, includeToJson: false)
   $ActorViewerCopyWith<ActorViewer> get copyWith =>
       throw _privateConstructorUsedError;
 }
@@ -52,10 +57,12 @@ abstract class $ActorViewerCopyWith<$Res> {
       ListViewBasic? blockingByList,
       @AtUriConverter() AtUri? blocking,
       @AtUriConverter() AtUri? following,
-      @AtUriConverter() AtUri? followedBy});
+      @AtUriConverter() AtUri? followedBy,
+      KnownFollowers? knownFollowers});
 
   $ListViewBasicCopyWith<$Res>? get mutedByList;
   $ListViewBasicCopyWith<$Res>? get blockingByList;
+  $KnownFollowersCopyWith<$Res>? get knownFollowers;
 }
 
 /// @nodoc
@@ -68,6 +75,8 @@ class _$ActorViewerCopyWithImpl<$Res, $Val extends ActorViewer>
   // ignore: unused_field
   final $Res Function($Val) _then;
 
+  /// Create a copy of ActorViewer
+  /// with the given fields replaced by the non-null parameter values.
   @pragma('vm:prefer-inline')
   @override
   $Res call({
@@ -78,6 +87,7 @@ class _$ActorViewerCopyWithImpl<$Res, $Val extends ActorViewer>
     Object? blocking = freezed,
     Object? following = freezed,
     Object? followedBy = freezed,
+    Object? knownFollowers = freezed,
   }) {
     return _then(_value.copyWith(
       isMuted: null == isMuted
@@ -108,9 +118,15 @@ class _$ActorViewerCopyWithImpl<$Res, $Val extends ActorViewer>
           ? _value.followedBy
           : followedBy // ignore: cast_nullable_to_non_nullable
               as AtUri?,
+      knownFollowers: freezed == knownFollowers
+          ? _value.knownFollowers
+          : knownFollowers // ignore: cast_nullable_to_non_nullable
+              as KnownFollowers?,
     ) as $Val);
   }
 
+  /// Create a copy of ActorViewer
+  /// with the given fields replaced by the non-null parameter values.
   @override
   @pragma('vm:prefer-inline')
   $ListViewBasicCopyWith<$Res>? get mutedByList {
@@ -123,6 +139,8 @@ class _$ActorViewerCopyWithImpl<$Res, $Val extends ActorViewer>
     });
   }
 
+  /// Create a copy of ActorViewer
+  /// with the given fields replaced by the non-null parameter values.
   @override
   @pragma('vm:prefer-inline')
   $ListViewBasicCopyWith<$Res>? get blockingByList {
@@ -132,6 +150,20 @@ class _$ActorViewerCopyWithImpl<$Res, $Val extends ActorViewer>
 
     return $ListViewBasicCopyWith<$Res>(_value.blockingByList!, (value) {
       return _then(_value.copyWith(blockingByList: value) as $Val);
+    });
+  }
+
+  /// Create a copy of ActorViewer
+  /// with the given fields replaced by the non-null parameter values.
+  @override
+  @pragma('vm:prefer-inline')
+  $KnownFollowersCopyWith<$Res>? get knownFollowers {
+    if (_value.knownFollowers == null) {
+      return null;
+    }
+
+    return $KnownFollowersCopyWith<$Res>(_value.knownFollowers!, (value) {
+      return _then(_value.copyWith(knownFollowers: value) as $Val);
     });
   }
 }
@@ -151,12 +183,15 @@ abstract class _$$ActorViewerImplCopyWith<$Res>
       ListViewBasic? blockingByList,
       @AtUriConverter() AtUri? blocking,
       @AtUriConverter() AtUri? following,
-      @AtUriConverter() AtUri? followedBy});
+      @AtUriConverter() AtUri? followedBy,
+      KnownFollowers? knownFollowers});
 
   @override
   $ListViewBasicCopyWith<$Res>? get mutedByList;
   @override
   $ListViewBasicCopyWith<$Res>? get blockingByList;
+  @override
+  $KnownFollowersCopyWith<$Res>? get knownFollowers;
 }
 
 /// @nodoc
@@ -167,6 +202,8 @@ class __$$ActorViewerImplCopyWithImpl<$Res>
       _$ActorViewerImpl _value, $Res Function(_$ActorViewerImpl) _then)
       : super(_value, _then);
 
+  /// Create a copy of ActorViewer
+  /// with the given fields replaced by the non-null parameter values.
   @pragma('vm:prefer-inline')
   @override
   $Res call({
@@ -177,6 +214,7 @@ class __$$ActorViewerImplCopyWithImpl<$Res>
     Object? blocking = freezed,
     Object? following = freezed,
     Object? followedBy = freezed,
+    Object? knownFollowers = freezed,
   }) {
     return _then(_$ActorViewerImpl(
       isMuted: null == isMuted
@@ -207,6 +245,10 @@ class __$$ActorViewerImplCopyWithImpl<$Res>
           ? _value.followedBy
           : followedBy // ignore: cast_nullable_to_non_nullable
               as AtUri?,
+      knownFollowers: freezed == knownFollowers
+          ? _value.knownFollowers
+          : knownFollowers // ignore: cast_nullable_to_non_nullable
+              as KnownFollowers?,
     ));
   }
 }
@@ -222,7 +264,8 @@ class _$ActorViewerImpl extends _ActorViewer {
       this.blockingByList,
       @AtUriConverter() this.blocking,
       @AtUriConverter() this.following,
-      @AtUriConverter() this.followedBy})
+      @AtUriConverter() this.followedBy,
+      this.knownFollowers})
       : super._();
 
   factory _$ActorViewerImpl.fromJson(Map<String, dynamic> json) =>
@@ -247,10 +290,12 @@ class _$ActorViewerImpl extends _ActorViewer {
   @override
   @AtUriConverter()
   final AtUri? followedBy;
+  @override
+  final KnownFollowers? knownFollowers;
 
   @override
   String toString() {
-    return 'ActorViewer(isMuted: $isMuted, isBlockedBy: $isBlockedBy, mutedByList: $mutedByList, blockingByList: $blockingByList, blocking: $blocking, following: $following, followedBy: $followedBy)';
+    return 'ActorViewer(isMuted: $isMuted, isBlockedBy: $isBlockedBy, mutedByList: $mutedByList, blockingByList: $blockingByList, blocking: $blocking, following: $following, followedBy: $followedBy, knownFollowers: $knownFollowers)';
   }
 
   @override
@@ -270,15 +315,27 @@ class _$ActorViewerImpl extends _ActorViewer {
             (identical(other.following, following) ||
                 other.following == following) &&
             (identical(other.followedBy, followedBy) ||
-                other.followedBy == followedBy));
+                other.followedBy == followedBy) &&
+            (identical(other.knownFollowers, knownFollowers) ||
+                other.knownFollowers == knownFollowers));
   }
 
-  @JsonKey(ignore: true)
+  @JsonKey(includeFromJson: false, includeToJson: false)
   @override
-  int get hashCode => Object.hash(runtimeType, isMuted, isBlockedBy,
-      mutedByList, blockingByList, blocking, following, followedBy);
+  int get hashCode => Object.hash(
+      runtimeType,
+      isMuted,
+      isBlockedBy,
+      mutedByList,
+      blockingByList,
+      blocking,
+      following,
+      followedBy,
+      knownFollowers);
 
-  @JsonKey(ignore: true)
+  /// Create a copy of ActorViewer
+  /// with the given fields replaced by the non-null parameter values.
+  @JsonKey(includeFromJson: false, includeToJson: false)
   @override
   @pragma('vm:prefer-inline')
   _$$ActorViewerImplCopyWith<_$ActorViewerImpl> get copyWith =>
@@ -300,7 +357,8 @@ abstract class _ActorViewer extends ActorViewer {
       final ListViewBasic? blockingByList,
       @AtUriConverter() final AtUri? blocking,
       @AtUriConverter() final AtUri? following,
-      @AtUriConverter() final AtUri? followedBy}) = _$ActorViewerImpl;
+      @AtUriConverter() final AtUri? followedBy,
+      final KnownFollowers? knownFollowers}) = _$ActorViewerImpl;
   const _ActorViewer._() : super._();
 
   factory _ActorViewer.fromJson(Map<String, dynamic> json) =
@@ -326,7 +384,12 @@ abstract class _ActorViewer extends ActorViewer {
   @AtUriConverter()
   AtUri? get followedBy;
   @override
-  @JsonKey(ignore: true)
+  KnownFollowers? get knownFollowers;
+
+  /// Create a copy of ActorViewer
+  /// with the given fields replaced by the non-null parameter values.
+  @override
+  @JsonKey(includeFromJson: false, includeToJson: false)
   _$$ActorViewerImplCopyWith<_$ActorViewerImpl> get copyWith =>
       throw _privateConstructorUsedError;
 }